1-7 working

+++ sync
@@ -70,6 +70,10 @@ def createSyncFile(inputDirectory):
 def updateSyncFile(inputDirectory):
     #load in json file, check if hash is different, check if date different
 
+    if (inputDirectory == firstDirectory):
+        directroyTwo = secondDirectory
+    else:
+        directroyTwo = firstDirectory
     with open(os.path.join(inputDirectory, '.sync')) as jsonData:
         jsonDict = json.load(jsonData)
 
@@ -82,8 +86,10 @@ def updateSyncFile(inputDirectory):
             if jsonDict[file][0][1] != "Deleted":
                 mod = time.strftime("%Y-%m-%d %H:%M:%S %z" , time.gmtime())
                 newDict[file].insert(0, (mod, "Deleted"))
+                if(os.path.isfile(os.path.join(directroyTwo,file))):
+                    pathToFile = os.path.join(directroyTwo, file)
+                    os.remove(pathToFile)
             continue
-
         newDigest = encodeFile(fullName)
 
         if not newDigest == jsonDict[file][0][1]:
@@ -113,13 +119,11 @@ def getDirectorySyncInfo(directory):
         return dirJson
 
 def addOldFileInfo(fileInfo,jsonDict, inputDirectory,fileName):
-    print("Add ofld file")
+
     date = fileInfo[0][0]
     digest = fileInfo[0][1]
     jsonDict[fileName]= []
     jsonDict[fileName].append((date,digest))
-    print("adding this date to the digest")
-    print(date)
     with open((os.path.join(inputDirectory, '.sync')), 'w') as outfile:
         json.dump(jsonDict, outfile,indent=4)
 
@@ -142,7 +146,7 @@ def checkAndCompareSyncFiles(directoryOne,directoryTwo):
         #if the file is not in the other directory, then add it and update that directories sync file WITH THE MATCHING DATE AND TIME
         #WORK FROM HERE
         #addOlfFileInfo
-        if file not in dirTwoJson and(os.path.isfile(fullName)):
+        if file not in dirTwoJson and(os.path.isfile(fullName)) and dirOneJson[file][0][1] != "Deleted":
             #needs to copy metadata
             shutil.copy2(fullName, directoryTwo)
             #just manually add it if it doesnt work
@@ -157,7 +161,7 @@ def checkAndCompareSyncFiles(directoryOne,directoryTwo):
             dirTwoDigest = fileTwo[1]
 
             #Digests are same, date is diff
-            if dirOneDigest == dirTwoDigest:
+            if dirOneDigest == dirTwoDigest and dirOneDigest != "Deleted" and dirTwoDigest != "Deleted":
                 #moved to checkifdatesdiff method
                 if fileOne[0] != fileTwo[0]:
                     mostRecentIsOne = getMostRecentDate(fileOne,fileTwo)
@@ -165,32 +169,21 @@ def checkAndCompareSyncFiles(directoryOne,directoryTwo):
                         addUpdatedDate(directoryTwo, file,dirTwoDigest,fileOne[0])
                     else:
                         addUpdatedDate(directoryOne, file,dirOneDigest,fileTwo[0])
-
-            elif(dirOneDigest == "Deleted"):
-                pathToFile = os.path.join(directoryTwo, file)
-                os.remove(pathToFile)
+            elif(dirOneDigest == "Deleted") and (os.path.isfile(os.path.join(directoryTwo, file))):
+                copyFileInfo(file,directoryOne,directoryTwo)
+                updateSyncFile(directoryOne)
+                updateSyncFile(directoryTwo)
             else:
-                print("digests are different")
-                print("file one date is ")
-                print(fileOne[0])
-                print("file two date is ")
-                print(fileTwo[0])
+
                 #digests are different, if dates are different get the most recent one and replace the other file
                 mostRecentIsOne = getMostRecentDate(fileOne,fileTwo)
                 if(mostRecentIsOne):
-                    print("most recent is one")
-                    print(file)
                     copyFileInfo(file,directoryOne,directoryTwo)
                     updateSyncFile(directoryOne)
                     updateSyncFile(directoryTwo)
                     #the first file is newest, so replace the contents of file two with file one
                 else:
-                    print("most recent is two")
-                    print(file)
-                    print(fileOne)
-                    print(fileTwo)
                     copyFileInfo(file,directoryTwo,directoryOne)
-
                     updateSyncFile(directoryOne)
                     updateSyncFile(directoryTwo)
 
@@ -200,109 +193,11 @@ def compDir():
     checkAndCompareSyncFiles(firstDirectory,secondDirectory)
     checkAndCompareSyncFiles(secondDirectory,firstDirectory)
 
-
-def compareDirectories():
-    updateSyncFile(firstDirectory)
-    updateSyncFile(secondDirectory)
-
-    with open(os.path.join(firstDirectory, '.sync')) as jsonData:
-        dirOneJson = json.load(jsonData)
-
-    with open(os.path.join(secondDirectory, '.sync')) as jsonDataTwo:
-        dirTwoJson = json.load(jsonDataTwo)
-
-    for file in dirOneJson:
-        fullName = os.path.join(firstDirectory, file)
-        if file not in dirTwoJson and(os.path.isfile(fullName)):
-            shutil.copy2(fullName, secondDirectory)
-            #do manual add to the json file, adding in old date etc etc
-            updateSyncFile(secondDirectory)
-        else:
-            fileOne = dirOneJson[file][0]
-            fileTwo = dirTwoJson[file][0]
-        #file exists in d2, check date, if diff the, then hash. if diff and date
-            dirOneDigest = fileOne[1]
-            dirTwoDigest = fileTwo[1]
-
-            #Digests are same, date is diff
-            if dirOneDigest == dirTwoDigest:
-                if fileOne[0] != fileTwo[0]:
-                    mostRecentIsOne = getMostRecentDate(fileOne,fileTwo)
-                    if not(mostRecentIsOne):
-                        addUpdatedDate(secondDirectory, file,dirTwoDigest,fileOne[0])
-                    else:
-                        addUpdatedDate(firstDirectory, file,dirOneDigest,fileTwo[0])
-            elif(dirOneDigest == "Deleted"):
-                pathToFile = os.path.join(secondDirectory, file)
-                os.remove(pathToFile)
-            else:
-                #Files are different, if dates are different get the most recent one and replace the other file
-                mostRecentIsOne = getMostRecentDate(fileOne,fileTwo)
-                if(mostRecentIsOne):
-                    print("first most recent is one")
-                    print(fullName)
-                    copyFileInfo(file,secondDirectory,firstDirectory)
-                    updateSyncFile(firstDirectory)
-                    updateSyncFile(firstDirectory)
-                    #the first file is newest, so replace the contents of file two with file one
-                else:
-                    print(fullName)
-                    print("first most recent is two")
-                    copyFileInfo(file,firstDirectory,secondDirectory)
-                    updateSyncFile(secondDirectory)
-
-    for file in dirTwoJson:
-        fullName = os.path.join(secondDirectory, file)
-        if file not in dirOneJson:
-            shutil.copy(fullName, firstDirectory)
-            updateSyncFile(firstDirectory)
-        else:
-            fileOne = dirOneJson[file][0]
-            fileTwo = dirTwoJson[file][0]
-        #file exists in d1, check date, if diff the, then hash. if diff and date
-            dirOneDigest = fileOne[1]
-            dirTwoDigest = fileTwo[1]
-            #Digests are same, date is diff
-            if dirOneDigest == dirTwoDigest:
-                if fileOne[0] != fileTwo[0]:
-                    mostRecentIsOne = getMostRecentDate(fileOne,fileTwo)
-                    if not (mostRecentIsOne):
-                        addUpdatedDate(secondDirectory, file,dirTwoDigest,fileOne[0])
-                    else:
-                        addUpdatedDate(firstDirectory, file,dirOneDigest,fileTwo[0])
-            elif(dirTwoDigest == "Deleted"):
-                pathToFile = os.path.join(firstDirectory, file)
-                os.remove(pathToFile)
-            else:
-                #Files are different, if dates are different get the most recent one and replace the other file
-                mostRecentIsOne = getMostRecentDate(fileOne,fileTwo)
-                if(mostRecentIsOne):
-                    print("sec most recent is one")
-                    print(fullName)
-                    copyFileInfo(file,secondDirectory,firstDirectory)
-
-                    updateSyncFile(firstDirectory)
-                    #the first file is newest, so replace the contents of file two with file one
-                else:
-                        print("sec most recent is else")
-                        print(fullName)
-                        copyFileInfo(file,firstDirectory,secondDirectory)
-                        updateSyncFile(secondDirectory)
-
-
 def copyFileInfo(inputFile,inputDirectory,outputDirectory):
     readFromFile = os.path.join(inputDirectory, inputFile)
     addInfoFile = os.path.join(outputDirectory, inputFile)
-    print("copying from to")
-    print(readFromFile)
-    print(outputDirectory)
-    shutil.copy2(readFromFile, outputDirectory)
-
-    '''if (os.path.isfile(readFromFile)):
-        if(os.path.isfile(addInfoFile)):
-            with open(readFromFile, 'rb') as fsrc:
-                with open(addInfoFile, 'wb') as fdest:
-                    shutil.copyfileobj(fsrc, fdest, 1000)'''
+    if(os.path.isfile(readFromFile)):
+        shutil.copy2(readFromFile, outputDirectory)
 
 def addUpdatedDate(inputDirectory,inputFile,digest,newDate):
     with open(os.path.join(inputDirectory, '.sync')) as jsonData:
